Module 1
Introduce the overall application - both the serverside components and
the CLI utility to interact with it. This module focuses on illustrating
proper environment configuration (rather than relying on the system
environment or tools like phpdotenv).

Your task is to relocate a hard-coded user identification string to a
safe location within the system environment that won't otherwise be
accessible to visitors or potentially leaked if the system encounters a
crash.