Module 4
Now users can authenticate, and they can use the system securely thanks
to stronger authentication systems. However, we're still trusting that
the user is providing us with valid information - mostly their email
address and their note. In our CLI tool this isn't a _huge_ concern, but
if paired with a web app we might be providing malicious parties with an
injection route!

Your task is to leverage sanitization and validation filters to prevent
malicious data from making it into the database.